requirements.txt
numpy==1.23.0
scipy==1.8.1
svgpathtools==1.5.0
svgwrite==1.4.2
matplotlib
fastf1
pyyaml